Changing tabs to spaces. Read description
We need to choose a coding style to stick to, a mix of tabs and spaces
is bad. My vote is for spaces because it keeps everything in-line no
matter who's looking at it on whatever editor they're using. You can
configure your editor to automatically replace tabs with spaces, it
feels exactly the same to use. Set it to 4 spaces per tab.